remove pemasukan nominal dan periode

+++ application/views/pemasukan/index.php
@@ -24,32 +24,15 @@
                     <h5 class="card-title">Tambah Kategori Pemasukan</h5>
                 </div>
                 <div class="card-body">
-                    <div class="row">
-                        <div class="col-6">
-                            <form action="<?= base_url("pemasukan/simpan") ?>" method="POST">
-                                <div class="form-group">
-                                    <label for="">Nama Kategori Pemasukan</label>
-                                    <input type="text" name="nama_jenis_pemasukan" class="form-control">
-                                </div>
-                                <div class="form-group">
-                                    <label for="">Periode</label>
-                                    <select name="periode_pemasukan" class="form-control">
-                                        <option value="11">Satu Kali</option>
-                                        <option value="1">Setiap 1 Bulan</option>
-                                        <option value="6">Setiap 6 Bulan</option>
-                                        <option value="12">Setiap 1 Tahun</option>
-                                    </select>
-                                </div>
-                                <div class="form-group">
-                                    <label>Nominal</label>
-                                    <input type="number" class="form-control" placeholder="Nominal per periode" name="nominal">
-                                </div>
-                                <button type="submit" class="btn btn-primary mb-3">
-                                    <i class="fas fa-save"></i> Simpan
-                                </button>
-                            </form>
+                    <form action="<?= base_url("pemasukan/simpan") ?>" method="POST">
+                        <div class="form-group">
+                            <label for="">Nama Kategori Pemasukan</label>
+                            <input type="text" name="nama_jenis_pemasukan" placeholder="Masukkan Kategori Pemasukan" class="form-control">
                         </div>
-                    </div>
+                        <button type="submit" class="btn btn-primary mb-3">
+                            <i class="fas fa-save"></i> Simpan
+                        </button>
+                    </form>
                 </div>
             </div> <!-- End Card -->
         </div>
@@ -65,8 +48,6 @@
                             <tr>
                                 <th>No</th>
                                 <th>Nama Kategori Pemasukan</th>
-                                <th>Periode</th>
-                                <th>Nominal</th>
                                 <th>Opsi</th>
                             </tr>
                         </thead>
@@ -82,20 +63,6 @@
                                 <td><?= $no ?></td>
                                 <td><?= $k->nama_jenis_pemasukan ?></td>
                                 <td>
-                                    <?php if ($k->periode_pemasukan == 1): ?>
-                                        Setiap 1 Bulan
-                                    <?php elseif ($k->periode_pemasukan == 6): ?>
-                                        Setiap 6 Bulan
-                                    <?php elseif ($k->periode_pemasukan == 12): ?>
-                                        Setiap 1 Tahun
-                                    <?php else: ?>
-                                        Satu Kali
-                                    <?php endif ?>
-                                </td>
-                                <td>
-                                    Rp. <?= number_format($k->nominal, 0); ?>
-                                </td>
-                                <td>
                                     <!-- <a href="<?= base_url() ?>pemasukan/detail/<?= $k->id_jenis_pemasukan ?>">
                                         <button class="btn btn-success text-white">
                                             <i class="fas fa-eye"></i>
@@ -106,7 +73,7 @@
                                             <i class="fas fa-edit"></i>
                                         </button>
                                     </a>
-                                    <a href="<?= base_url() ?>pemasukan/hapus/<?= $k->id_jenis_pemasukan ?>">
+                                    <a href="<?= base_url() ?>pemasukan/hapus/<?= $k->id_jenis_pemasukan ?>" onclick="return confirm('Apakah anda ingin menghapus data ini ?')">
                                         <button class="btn btn-danger">
                                             <i class="fas fa-trash"></i>
                                         </button>
